requirements.txt
# HTTP & scraping
requests>=2.28.0
beautifulsoup4>=4.12.0

# Images (used in ingester)
Pillow>=10.0.0

# Embeddings & vector DB
sentence-transformers>=2.2.0
chromadb>=0.4.0

# LLM (OpenAI / Ollama)
litellm>=1.0.0
